requirements.txt
streamlit==1.31.0
langchain==0.1.16
langchain-community==0.0.34
llama-cpp-python==0.2.55
sentence-transformers==2.5.1
chromadb==0.4.24
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==0.8.11
pandas==2.2.1
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1
tqdm==4.66.2
requests==2.31.0
beautifulsoup4==4.12.3
transformers==4.39.3
torch==2.2.2
torchvision==0.17.2
plotly==5.19.0
altair==5.2.0
